mm_stem/requirements.txt
# MM-STEM Requirements
# Multimodal STEM (Scaling Transformers with Embedding Modules) Implementation

# Core dependencies
torch>=2.0.0
transformers>=4.36.0
tokenizers>=0.15.0
accelerate>=0.25.0

# Vision
Pillow>=9.0.0
torchvision>=0.15.0

# Data processing
datasets>=2.14.0
numpy>=1.24.0

# Training utilities
tqdm>=4.65.0
pyyaml>=6.0

# Logging and monitoring
tensorboard>=2.14.0
wandb>=0.15.0  # optional

# Development
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0

# Jupyter support (optional)
ipykernel>=6.25.0
jupyter>=1.0.0

# Memory efficiency (optional but recommended)
bitsandbytes>=0.41.0  # 8-bit optimizers
deepspeed>=0.10.0  # ZeRO optimization
